requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.8
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
loguru==0.7.2
slowapi==0.1.9
websockets==12.0

# Nuevas dependencias para integración
aiohttp==3.9.1
asyncio-timeout==4.0.3
pymongo==4.6.0

# Testing adicional
pytest==7.4.3
pytest-asyncio==0.21.1
structlog==23.2.0
